Use `DbContext`-specific initializer
