+++ src/Testing/LeanCode.IntegrationTestHelpers/LeanCodeTestFactory.cs
@@ -89,9 +89,6 @@ namespace LeanCode.IntegrationTestHelpers
                     // `DbContextsInitializer` successfully drop the database more frequently. :)
                     services.Configure<HostOptions>(
                             opts => opts.ShutdownTimeout = TimeSpan.FromSeconds(15));
-
-                    // Registering it here will make it the first `IHostedService` in the collection
-                    services.AddTransient<IHostedService, DbContextsInitializer>();
                 });
         }
